Move WO state-size Prom metrics into window::metrics.
Keep central observability for worker/transport gauges; WO metric names and
emit_window_operator_metrics live with the operator. Worker dispatches
TaskOperatorMetrics::Window to that helper.
